machine-learning/agents/langchain/modules/langfuse/pyproject.toml
[project]
name = "Demo"
version = "0.1.0"
description = "My reference project"
requires-python = "~=3.13.0"
dependencies = [
    "langchain~=1.0",
    "langfuse~=3.0",
    "langchain-google-genai~=4.0",
    "python-dotenv~=1.0",
]
